# Speedup CI: use playwright container instead of installing system deps