--- a/.changelog/1956.internal.md
+++ b/.changelog/1956.internal.md
@@ -1,0 +1,1 @@
+Speedup CI: use playwright container instead of installing system deps