Same as https://github.com/oasisprotocol/wallet/pull/2166/commits/eb3bb1d55ee313af4ece199170a9cad3b47488b0

Before:
 2m 21s
 2m 31s
 2m 34s
 4m 0s
 3m 19s

After:
 2m 51s
 2m 7s
 3m 13s
 2m 1s